backend/requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.12
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
narwhals==2.12.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
protobuf==6.33.1
psutil==7.1.3
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.29.0
safetensors==0.7.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.230
ultralytics-thop==2.0.18
urllib3==2.5.0
watchdog==6.0.0

